Saving data to database, getting data with token and session

+++ KryptoProject/crypto/models.py
@@ -4,8 +4,8 @@ from django.contrib.auth.models import User as mainUser
 
 class UserData(models.Model):
     username = models.CharField(max_length=50,unique=True)
-    wallet_dollars = models.IntegerField(default = 1000)
-    wallet_bit = models.IntegerField(default = 0)
+    wallet_dollars = models.FloatField(default = 1000.0)
+    wallet_bit = models.FloatField(default = 0.0)
 
     def __str__(self):
         return self.username
